Issue #200 #150 #149 Reporting/Self Monitoring
- Issue 200 check the filter, matcher, and router input channel depths.
If any are at capacity check for a slow matcher or filter and terminate the
sandbox if necessary.

- Issue 150 Message counts were added to the router and sandbox plugins.
They can be added to other plugins as desired using a ProcessMessageCount
field in their ReportMsg function.

- Issue 149 Recording the amount of time each plugin holds onto a message is
expensive (on my system ~40ns per timer) so instead only sampling is done and
restricted to the matcher and SandboxFilter where it is required to detect
problematic code.