Modify SeekInFile and FileHashMismatch to handle gzipped files
Previously, Seek was used on the current fd to seek to the
penultimate line in the file and hash the last line.

Where the current file is gzipped, consume bytes from the
gzip reader to give the same effect as Seek.

This creates unnecessary garbage but is probably the simplest
way to emulate Seek.