Only check for backpressure on dynamic sandbox plugins
On startup the system is slammed by all the plugins contending for resources
which can skew the backpressure sample timing by over 200x.  Occasionally this
causes static sandboxes to be terminated as 'running too slowly'.  Static
sandboxes should be considered part of the system, trusted and therefore
exempt from this check.