docs: better page titles for input plugins
The page titles for input plugins were *Input. These were showing up in
tocs and sidebars and weren't the most friendly of strings for humans to
read.

This commit converts the titles to something more human friendly and
adds a plugin name entry defining the name for the plugin to be used
in configs. This latter bit was necessary because the page titles were
formerly serving this role.

Having a dedicated RST entity to declare the plugin name is probably
appropriate. But the patch author was too lazy to perform this work.