Add disk-based buffering for ElasticSearch output
As discussed on the Heka mailinglist, this is a commit to introduce a
use_buffering mode to the ElasticSearch output plugin. It changes the
default behavior to buffer to disk, like the TCP Output plugin, with all
the other options of the TCP Output plugin and the additional option to
set use_buffering to false to disable it (this just dumps any failed
packets to nowhere, though - no retry logic).

This also buffers in "chunks" rather than individual messages, so there
still is the tiny possibility that if a message has been received and
hasn't filled the commit buffer yet, it can be lost if Heka dies, but the
default commit buffer is sent every 1 second so this state is a lot better
than quietly dropping messages if they fail to send even once.