Update handling of EOF in framed protobuf streams.
The code keeps track explicitly of whether we've reached the end
of the stream. We also continue returning any well-formed records
until EOF, so you should not need to use the GetRemainingData()
function unless you have partial message bytes at the end of the
stream and you want to do something with them.